[SofaSimulationCore] Move MechanicalVAvailVisitor to dedicated folder
